refactor: make utils great again
Changed:
- Replace `util.PubKeyToAddr` with `crypto.PubkeyToAddress`.
- Drop `ParseEndpointPort`.
- Move all network-related stuff to netutil package.
