+++ util/netutil/net.go
@@ -1,6 +1,7 @@
 package netutil
 
 import (
+	"errors"
 	"fmt"
 	"net"
 	"strconv"
@@ -60,4 +61,57 @@ func (m *TCPAddr) UnmarshalYAML(unmarshal func(interface{}) error) error {
 
 	m.TCPAddr = *tcpAddr
 	return nil
+}
+
+func GetAvailableIPs() (availableIPs []net.IP, err error) {
+	ifaces, err := net.Interfaces()
+	if err != nil {
+		return nil, err
+	}
+
+	for _, i := range ifaces {
+		addrs, err := i.Addrs()
+		if err != nil {
+			return nil, err
+		}
+		for _, addr := range addrs {
+			var ip net.IP
+			switch v := addr.(type) {
+			case *net.IPNet:
+				ip = v.IP
+			case *net.IPAddr:
+				ip = v.IP
+			}
+			if ip != nil && ip.IsGlobalUnicast() {
+				availableIPs = append(availableIPs, ip)
+			}
+		}
+	}
+	if len(availableIPs) == 0 {
+		return nil, errors.New("could not determine a single unicast addr, check networking")
+	}
+
+	return availableIPs, nil
+}
+
+func IsPrivateIP(ip net.IP) bool {
+	return isPrivateIPv4(ip) || isPrivateIPv6(ip)
+}
+
+func isPrivateIPv4(ip net.IP) bool {
+	_, private24BitBlock, _ := net.ParseCIDR("10.0.0.0/8")
+	_, private20BitBlock, _ := net.ParseCIDR("172.16.0.0/12")
+	_, private16BitBlock, _ := net.ParseCIDR("192.168.0.0/16")
+	return private24BitBlock.Contains(ip) ||
+		private20BitBlock.Contains(ip) ||
+		private16BitBlock.Contains(ip) ||
+		ip.IsLoopback() ||
+		ip.IsLinkLocalUnicast() ||
+		ip.IsLinkLocalMulticast()
+}
+
+func isPrivateIPv6(ip net.IP) bool {
+	_, block, _ := net.ParseCIDR("fc00::/7")
+
+	return block.Contains(ip) || ip.IsLoopback() || ip.IsLinkLocalUnicast() || ip.IsLinkLocalMulticast()
 }
